dump/requirements.txt
dateparser==1.1.1
jsonlines==3.0.0
kafka_python==2.0.2
